Adapt offer page to be able to show insurances that are not house/apartment
Use contract type to determine which quote bundle that is selected, and what is shown in the URL. This is because all car quote variations have the same quote type, but not same contract type. This also happens to work for other bundles.

Adapt SummaryDetails to only show information that's available with the current data.