Show different Start date label if car switcher: "Select date" instead of "When the old one expires"
This is because we cannot guarantee that a date that is within one week can be switched in time. So user's have to make a choice about the date, so that we can prompt them if the date is too soon.

If it would just say "when the old one expires" users might not interact with the element and then we couldn't show the warning notice either.