Introduce a useHasScrolledPastElement hook and use that to set positioning for Sidebar
There's a concern that we might have a too long Sidebar now with the Car switching content, and that users might not be able to reach the "Continue to checkout" button. This commit sets the Sidebar to absolutely positioned by default instead of fixed, and then sets fixed to the bottom of the screen when scrolled past. For desktop, to emulate the fixed positioning we set fixed to top when scrolled past instead.