Use renewal date from Insurely as suggested start date (#1185)
## What?

Use renewal date for the old insurance from Insurely data collection to suggest start date on Offer page.

Add new reusable `useExternalDataCollection` hook. Refactor `InsuranceSelector` to use new hook.

Refactor `StartDate` component and add button to update start date to match renewal date of old insurance.

## Why?

We want to help people switch and BankSignering requires us to send a date for the start of the new insurance.

This only works for Car now but when we can match exposure for other insurance types we can enable this same feature.

There's no way to set a default date in our date picker -- that's why I went with a button the user has to click. This is also to inform the user where the date comes from. I made the button disabled if the start date already matches the renewal date to highlight the same thing.

**Ticket(s): []**