Refactor how to populate start date from Insurely (#1190)
## What?

Refactor how we populate start date from Insurely for Car on offer page.

Re-use the same solution inside `DateInput` but make it possible to set a defined start date or `null`.

## Why?

It was confusing with 2 different solutions/labels.
This solution is not very scalable so if we do additional changes here we should consider a bigger refactoring.