Wait for response before dismissing edit modal (#1218)
## What?

Wait before closing edit modal so all mutations respond first.

Update error message when household size field value is invalid.

## Why?

We were reading a value outside the function before. This encloses the variable in the function scope so we are actually reading the value from the previous mutation 🤦🏻 Now we return the response and read it in the same context so they are always in sync.

I changed the error message as well since before it said "wrong format".

Thanks to @fredriklagerblad for reporting the issue!

**Ticket(s): []**